Correct font data decoding
My assumptions about doubling up on columns were all wrong.  I knew from
YouTube footage of the PSX version, and badly-compressed screenshots of
the original, that the row width in bytes was half the graphic width in
pixels.  However, firing up the original in an emulator, there was no
column doubling: the font really did have double the horizontal
resolution compared to what I had decoded.  Turns out the font data is
actually 4bpp, so packed 2 pixels per byte, not 8bpp as I had assumed.

Also, render level names in the correct place on screen.

This was measured pixel-for-pixel by taking screenshots of the original
in an emulator, and comparing them against my version.  In the process,
I discovered that the original - when running in mode 28, anyway -
actually chops off the bottom-most & right-most rows of its glyphs!

Signed-off-by: Philip Allison <mangobrain@googlemail.com>