Count time between frames manually
Instead of using SDL_gfx and its "framerate manager", count time between
frame updates manually, and pass it in to game object state update code.
Also attempt to detect whether or not vsync is enabled, and add an
explicit delay to the main loop if it is not.

Note that the game object state update code doesn't actually *use* the
elapsed time yet, but at least it's passed in.

Signed-off-by: Philip Allison <philip.allison@smoothwall.net>