Split main game loop code into a GameLoop class
Instead of implementing input handling & game state update directly in
the main loop, implement an abstract GameLoop class, with functionality
for indicating when to switch to a new GameLoop.  Implement the "InGame"
loop derived from this, and move the aforementioned input handling and
game state update into its update function.

There are still a few rough edges, in particular the way GameLoop
switching is handled (there is an overlap during which the current and
next GameLoops both exist, which could be wasteful of resources), and
the usage of the "objects_left" global variable.

Coming soon: GameLoop-derived classes for the main menu & pause menu.

Signed-off-by: Philip Allison <mangobrain@googlemail.com>