Fix processing of already-pressed keys in Menu
In the Menu class, don't fire events based on keys which were already
pressed when the Menu was created.  This will help stop, for example,
the pause menu from going straight into a new game when selecting "Quit"
(because that will go back to the main menu, with "Start Game" selected
and the relevant key already pressed).

Signed-off-by: Philip Allison <mangobrain@googlemail.com>