Rethink generation & distrib of icons & .desktop
Generate the various fixed-size icons as BUILT_SOURCES, so that we can
distribute them via dist_noinst_DATA rather than a dist-hook, which is
cleaner.  Also put them in $(srcdir) explicitly when generated, so that
VPATH builds from a distribution tarball (where the files *will* be in
$(srcdir), not the current working directory) don't run into issues.

Delete the .desktop file as part of clean-local, otherwise it isn't
cleaned up at all.

Signed-off-by: Philip Allison <mangobrain@googlemail.com>